% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(
  request_url = "https://apis.vinmonopolet.no/products/v0/details-normal",
  primary_key = "c8971627fe1d48b7bbb336a7ab924660"
)
}
\arguments{
\item{primary_key}{}
}
\value{

}
\description{
Get data
}
